requirements.txt
anthropic>=0.40.0
edge-tts>=6.1.9
elevenlabs>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
srt>=3.5.3
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
